data function must return an object data defines which data will be used in the VueJS app mount defines which part of the DOM will be manipulated by the Vue
v-model establish a connection between the input and the value in the data property (it listens to the value change and automatically update the value)
v-on is used to add event listener (e.g. v-on:click)
v-for is used to replicate html element (e.g. <li v-for="goal in goals">{{ goal }}</li>)